requirements.txt
Flask
google-cloud-storage
google-generativeai
functions-framework
pdfplumber
python-dotenv
